Add upgrade::apply
